+++ lib/gif_view.dart
@@ -1,6 +1,6 @@
+library gif_view;
+
 import 'dart:async';
-// ignore: unnecessary_import
-import 'dart:typed_data';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -13,41 +13,58 @@ import 'src/gif_loader.dart';
 export 'package:gif_view/src/gif_cache_provider.dart';
 export 'package:gif_view/src/gif_controller.dart';
 
-///
-/// Created by
-///
-/// ─▄▀─▄▀
-/// ──▀──▀
-/// █▀▀▀▀▀█▄
-/// █░░░░░█─█
-/// ▀▄▄▄▄▄▀▀
-///
-/// Rafaelbarbosatec
-/// on 23/09/21
-
 class GifView extends StatefulWidget {
+  /// Optional controller to manage GIF playback externally.
   final GifController? controller;
+
+  /// Duration between frames in the GIF animation.
+  /// Defaults is the fps original file
   final int? frameRate;
   final ImageProvider image;
+
+  /// The height of the GIF view widget.
+  /// If null, the widget will use its parent's height constraints.
   final double? height;
+
+  /// The width of the GIF view widget.
+  /// If null, the widget will use its parent's width constraints.
   final double? width;
+  final Widget Function(BuildContext context, Exception error)? errorBuilder;
   final WidgetBuilder? progressBuilder;
+
+  /// How to fit the image within its bounds.
+  /// Uses Flutter's BoxFit enum.
+  /// Defaults to BoxFit.contain.
   final BoxFit? fit;
   final Color? color;
   final BlendMode? colorBlendMode;
   final AlignmentGeometry alignment;
-  final ImageRepeat repeat;
+  final ImageRepeat imageRepeat;
   final Rect? centerSlice;
   final bool matchTextDirection;
   final bool invertColors;
-  final bool withOpacityAnimation;
   final FilterQuality filterQuality;
   final bool isAntiAlias;
-  final Widget Function(BuildContext context, Exception error)? errorBuilder;
+
+  final bool withOpacityAnimation;
   final Duration? fadeDuration;
+
+  /// Controls whether the GIF animation should auto-start.
+  /// Defaults to true.
+  final bool autoPlay;
+
+  /// Determines if the GIF should loop continuously.
+  /// When false, the animation will play once and stop.
+  /// Defaults to true.
+  final bool loop;
+  final bool playInverted;
   final void Function()? onStart;
+
+  /// Callback function that triggers when animation completes.
+  /// Only called when [loop] is false.
   final void Function()? onFinish;
   final void Function(int frame)? onFrame;
+  final void Function(int totalFrames)? onLoaded;
 
   GifView.network(
     String url, {
@@ -61,7 +78,7 @@ class GifView extends StatefulWidget {
     this.color,
     this.colorBlendMode,
     this.alignment = Alignment.center,
-    this.repeat = ImageRepeat.noRepeat,
+    this.imageRepeat = ImageRepeat.noRepeat,
     this.centerSlice,
     this.matchTextDirection = false,
     this.invertColors = false,
@@ -75,6 +92,10 @@ class GifView extends StatefulWidget {
     this.onStart,
     this.onFinish,
     this.onFrame,
+    this.onLoaded,
+    this.autoPlay = true,
+    this.loop = true,
+    this.playInverted = false,
   })  : image = NetworkImage(url, scale: scale, headers: headers),
         super(key: key);
 
@@ -90,7 +111,7 @@ class GifView extends StatefulWidget {
     this.color,
     this.colorBlendMode,
     this.alignment = Alignment.center,
-    this.repeat = ImageRepeat.noRepeat,
+    this.imageRepeat = ImageRepeat.noRepeat,
     this.centerSlice,
     this.matchTextDirection = false,
     this.invertColors = false,
@@ -104,6 +125,10 @@ class GifView extends StatefulWidget {
     this.onStart,
     this.onFinish,
     this.onFrame,
+    this.onLoaded,
+    this.autoPlay = true,
+    this.loop = true,
+    this.playInverted = false,
   })  : image = AssetImage(asset, package: package, bundle: bundle),
         super(key: key);
 
@@ -111,7 +136,7 @@ class GifView extends StatefulWidget {
     Uint8List bytes, {
     Key? key,
     this.controller,
-    this.frameRate = 15,
+    this.frameRate,
     this.height,
     this.width,
     this.progressBuilder,
@@ -119,7 +144,7 @@ class GifView extends StatefulWidget {
     this.color,
     this.colorBlendMode,
     this.alignment = Alignment.center,
-    this.repeat = ImageRepeat.noRepeat,
+    this.imageRepeat = ImageRepeat.noRepeat,
     this.centerSlice,
     this.matchTextDirection = false,
     this.invertColors = false,
@@ -132,6 +157,10 @@ class GifView extends StatefulWidget {
     this.onStart,
     this.onFinish,
     this.onFrame,
+    this.onLoaded,
+    this.autoPlay = true,
+    this.loop = true,
+    this.playInverted = false,
   })  : image = MemoryImage(bytes, scale: scale),
         super(key: key);
 
@@ -139,7 +168,7 @@ class GifView extends StatefulWidget {
     Key? key,
     required this.image,
     this.controller,
-    this.frameRate = 15,
+    this.frameRate,
     this.height,
     this.width,
     this.progressBuilder,
@@ -147,7 +176,7 @@ class GifView extends StatefulWidget {
     this.color,
     this.colorBlendMode,
     this.alignment = Alignment.center,
-    this.repeat = ImageRepeat.noRepeat,
+    this.imageRepeat = ImageRepeat.noRepeat,
     this.centerSlice,
     this.matchTextDirection = false,
     this.invertColors = false,
@@ -159,13 +188,38 @@ class GifView extends StatefulWidget {
     this.onStart,
     this.onFinish,
     this.onFrame,
+    this.onLoaded,
+    this.autoPlay = true,
+    this.loop = true,
+    this.playInverted = false,
   }) : super(key: key);
 
-  static Future<void> preFetch(ImageProvider image) async {
+  /// Pre-fetches GIF images and stores them in cache for faster loading.
+  ///
+  /// [image] A ImageProvider to be pre-fetched.
+  ///
+  /// Example:
+  /// ```dart
+  /// await GifView.preFetch(AssetImage());
+  /// ```
+  static Future<void> preFetchImage(ImageProvider image) async {
     await GifLoader.instance.fetch(image);
   }
 
-  static void setCacheProvider(GifCacheProvider provider) {
+  static Future<void> clearCache() async {
+    await GifLoader.instance.clearCache();
+  }
+
+  /// Sets a custom cache provider for GIF image storage.
+  ///
+  /// [provider] An implementation of [GifCacheProvider] to handle caching.
+  /// If null, reverts to the default cache provider.
+  ///
+  /// Example:
+  /// ```dart
+  /// GifView.setCacheProvider(MyCustomCacheProvider());
+  /// ```
+  static void setCacheProvider(GifCacheProvider? provider) {
     GifLoader.instance.setCacheProvider(provider);
   }
 
@@ -187,6 +241,14 @@ class GifViewState extends State<GifView> with SingleTickerProviderStateMixin {
       );
     }
     controller = widget.controller ?? GifController();
+    controller.init(
+      autoPlay: widget.autoPlay,
+      inverted: widget.playInverted,
+      loop: widget.loop,
+      onStart: widget.onStart,
+      onFinish: widget.onFinish,
+      onFrame: widget.onFrame,
+    );
     controller.addListener(_listener);
     WidgetsBinding.instance.addPostFrameCallback(
       (_) => _loadImage(),
@@ -251,7 +313,7 @@ class GifViewState extends State<GifView> with SingleTickerProviderStateMixin {
       color: widget.color,
       colorBlendMode: widget.colorBlendMode,
       alignment: widget.alignment,
-      repeat: widget.repeat,
+      repeat: widget.imageRepeat,
       centerSlice: widget.centerSlice,
       matchTextDirection: widget.matchTextDirection,
       invertColors: widget.invertColors,
@@ -271,8 +333,11 @@ class GifViewState extends State<GifView> with SingleTickerProviderStateMixin {
           frameRate: widget.frameRate,
         ).build();
         if (frames.isNotEmpty) {
+          widget.onLoaded?.call(frames.length);
           controller.configure(frames, updateFrames: updateFrames);
           _animationController?.forward(from: 0);
+        } else {
+          controller.error(Exception('Can not load image'));
         }
       } else {
         controller.error(Exception('Can not load image'));