+++ lib/src/gif_controller.dart
@@ -9,25 +9,33 @@ class GifController extends ChangeNotifier {
   GifStatus status = GifStatus.loading;
   Exception? exception;
 
-  final bool autoPlay;
-  final VoidCallback? onFinish;
-  final VoidCallback? onStart;
-  final ValueChanged<int>? onFrame;
+  VoidCallback? _onFinish;
+  VoidCallback? _onStart;
+  ValueChanged<int>? _onFrame;
 
-  bool loop;
-  bool _inverted;
+  bool _autoPlay = true;
+  bool _loop = true;
+  bool _inverted = false;
+
   int get index => _currentIndex;
 
   bool _isDisposed = false;
 
-  GifController({
-    this.autoPlay = true,
-    this.loop = true,
+  void init({
+    bool autoPlay = true,
+    bool loop = true,
     bool inverted = false,
-    this.onStart,
-    this.onFinish,
-    this.onFrame,
-  }) : _inverted = inverted;
+    VoidCallback? onStart,
+    VoidCallback? onFinish,
+    ValueChanged<int>? onFrame,
+  }) {
+    _autoPlay = autoPlay;
+    _loop = loop;
+    _inverted = inverted;
+    _onStart = onStart;
+    _onFinish = onFinish;
+    _onFrame = onFrame;
+  }
 
   void _run() {
     switch (status) {
@@ -36,7 +44,6 @@ class GifController extends ChangeNotifier {
         _runNextFrame();
         break;
       case GifStatus.stoped:
-        onFinish?.call();
         _currentIndex = 0;
         break;
       case GifStatus.loading:
@@ -59,23 +66,25 @@ class GifController extends ChangeNotifier {
       if (_currentIndex > 0) {
         int newIndex = _currentIndex - 1;
         _currentIndex = (newIndex % _frames.length);
-      } else if (loop) {
+      } else if (_loop) {
         _currentIndex = _frames.length - 1;
       } else {
+        _onFinish?.call();
         status = GifStatus.stoped;
       }
     } else {
       if (_currentIndex < _frames.length - 1) {
         int newIndex = _currentIndex + 1;
         _currentIndex = (newIndex % _frames.length);
-      } else if (loop) {
+      } else if (_loop) {
         _currentIndex = 0;
       } else {
+        _onFinish?.call();
         status = GifStatus.stoped;
       }
     }
 
-    onFrame?.call(_currentIndex);
+    _onFrame?.call(_currentIndex);
     notifyListeners();
     _run();
   }
@@ -102,7 +111,7 @@ class GifController extends ChangeNotifier {
       } else {
         _currentIndex = isReversing ? _frames.length - 1 : _currentIndex;
       }
-      onStart?.call();
+      _onStart?.call();
       _run();
     } else {
       status = _inverted ? GifStatus.reversing : GifStatus.playing;
@@ -124,7 +133,7 @@ class GifController extends ChangeNotifier {
   }
 
   void seek(int index) {
-    if (_frames.isEmpty) return;
+    if (_frames.isEmpty || _isDisposed) return;
     _currentIndex = (index % _frames.length);
     notifyListeners();
   }
@@ -134,7 +143,7 @@ class GifController extends ChangeNotifier {
     _frames = frames;
     if (!updateFrames || status == GifStatus.loading) {
       status = GifStatus.stoped;
-      if (autoPlay) {
+      if (_autoPlay) {
         play();
       }
       notifyListeners();
@@ -142,12 +151,18 @@ class GifController extends ChangeNotifier {
   }
 
   void error(Exception e) {
+    if (_isDisposed) {
+      return;
+    }
     exception = e;
     status = GifStatus.error;
     notifyListeners();
   }
 
   void loading() {
+    if (_isDisposed) {
+      return;
+    }
     status = GifStatus.loading;
     notifyListeners();
   }